+++ changelog.md
@@ -33,6 +33,7 @@ Codion Change Log
 - FrameworkMessages.insert() added.
 - EntityTablePanel constructor overloaded with edit panel parameter, adds Add and Edit controls when one is provided.
 - EntityEditPanel.KEYBOARD_SHORTCUTS added.
+- TabbedApplicationLayout.applicationPanel() added.
 
 ## 0.17.35
 ### is.codion.common.model